Implement spec-06: CI, ad-hoc signed release workflow & accurate README
- GitHub Actions CI on macOS: pnpm check, frontend tests, clippy -D warnings,
  cargo test, full tauri build (read-only token)
- Release workflow on v* tags: universal macOS build, ad-hoc signed
  (signingIdentity "-"), zipped .app attached to a draft GitHub release
- docs/release-checklist.md: RC gates, manual app matrix, clean-Mac smoke
  run, shipping steps, privacy check
- README rewritten to shipped behavior: automatic capture with clipboard
  fallback, no diff view in v1, install & one-time Gatekeeper approval,
  privacy section, building from source, CI badge
- Fix four pre-existing clippy lints so the new -D warnings gate passes

+++ src-tauri/src/platform/macos.rs
@@ -84,7 +84,7 @@ impl SelectionBackend for MacosSelectionBackend {
         let app = workspace.frontmostApplication()?;
         Some(SourceApp {
             bundle_id: app.bundleIdentifier().map(|s| s.to_string()),
-            pid: app.processIdentifier() as i32,
+            pid: app.processIdentifier(),
             name: app.localizedName().map(|s| s.to_string()),
         })
     }
